requirements.txt
configparser>=3.5.0
paramiko-ng>=2.5.0
pytz>=2017.3
tzlocal==2.0.0
requests==2.27.1
pyTelegramBotAPI>=3.6.3 
networkx>=2.6
matplotlib>=2.1.2
future>=0.13.1
jinja2>=2.10.1
slack-sdk>=3.4.0
peewee>=3.14.10
PyMySQL>=1.0.2
distro>=1.2.0
bottle>=0.12.20
psutil>=5.9.1
pdpyras>=4.5.2
pika>=1.3.1
Werkzeug==3.0.3
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Pydantic==0.12.0
flask-swagger==0.2.14
ansible-core>=2.11.12
ansible-runner==2.3.1
python-whois>=0.8.0
netaddr>=0.10.1
python-ldap>=3.3.1
